requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.4
django-channels==0.7.0
django-extensions==3.0.3
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.4
python-decouple==3.3
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
psycopg2-binary==2.8.4
channels==2.4.0
channels-redis==2.4.2
